requirements.txt
opencv-python
mtcnn
keras==3.6.0
keras-facenet==0.3.2
numpy==1.26.4
tensorflow==2.18.0
scikit-learn==1.5.2
joblib
mysql-connector-python
ultralytics